+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.03...master)
 
 * `Select::first`: apply default value
+* `Database\DB`: Custom engine (mysql, ...)
 
 ## [v0.0.03](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.02...v0.0.03) - 8/29/2024